审计修复与安全加固（鉴权/迁移/后台回归/部署）+ 文档同步
- 认证：登录同时下发 HttpOnly access_token/refresh_token Cookie（XSS 纵深防御），
  Secure 标志按请求协议判定（https 才置），不再依赖 APP_ENV；登出由后端清除 Cookie；
  后端 get_current_user 同时接受 Bearer 与 access_token Cookie 两种凭证。
- 迁移 4_：t_*_category.sort_order 改 DOUBLE PRECISION、t_news.status 默认 DRAFT、
  清理 WP 迁移残留表 t_migration_record/batch、新增 t_setting 表。
- 部署：移除 Docker（Dockerfile/Dockerfile.pg/docker-compose.yml/.dockerignore/docker/），
  改为 uv 虚拟环境 + 1Panel 部署（PG/Redis 由 1Panel 管理）。
- 管理后台 admin-next：middleware 守卫改读后端 access_token Cookie、移除明文 cookie 写入；
  修复两处致全站 500 的回归（ToastContext TDZ、categories 页 Modal 具名导入）。
- 文档：同步鉴权 Cookie 双重通道、迁移、部署说明；修正 F-08 描述、清理 admin_token 误标。

+++ backend/product/schemas.py
@@ -109,7 +109,7 @@ class CategoryVO(BaseModel):
     id: int
     name: str
     slug: str
-    sort_order: int = 0
+    sort_order: float = 0.0
 
     @classmethod
     def from_model(cls, m) -> CategoryVO:  # type: ignore[valid-type]
@@ -125,7 +125,7 @@ class CategoryCreate(BaseModel):
 
     name: str = Field(..., max_length=100)
     slug: str = Field(..., max_length=100)
-    sort_order: int | None = None  # 缺省时落到末尾
+    sort_order: float | None = None  # 缺省时落到末尾
 
 
 class CategoryUpdate(BaseModel):
@@ -133,7 +133,7 @@ class CategoryUpdate(BaseModel):
 
     name: str | None = None
     slug: str | None = None
-    sort_order: int | None = None
+    sort_order: float | None = None
 
 
 class ReorderReq(BaseModel):
